Fix the nightly bot stop/restart: boot cost, blast pacing, finalize lock
Three independent problems made the bot go quiet for a minute or two at
peak (evening IST) hours, plus two bugs the same logs exposed.

Boot cost — every restart was ~75s of total downtime, because init_db
took 71s before polling even started. _migrate_add_columns issued a
connect + BEGIN + ALTER + COMMIT per column, ~110 of them, against a
remote Postgres. It now reads the live schema in one query and skips
columns that already exist, runs the remaining DDL batches on a single
connection with a SAVEPOINT each, and gates those batches behind a
signature marker so an applied batch costs nothing on later boots. The
signature is a hash of the statements, so editing them re-runs the batch
automatically. Measured against a local Postgres with 30ms of simulated
latency: 660 round trips -> 125, 21.1s -> 4.2s. Seeds that queried once
per row now query once per batch.

Notification blasts — fire_schedule sent to every target user in a tight
unpaced loop on the bot's event loop, holding one DB session open for the
whole run. Sends are now paced under Telegram's global rate limit, users
are snapshotted so no session is pinned across the blast, delivery rows
are written in chunks on worker threads, and the blast is detached from
the heartbeat rather than awaited (awaiting it made the scheduler skip
every 30s heartbeat for the length of the broadcast). last_fired_at is
committed before the first message: it used to be written at the end, so
the next tick would pick the same schedule up and fire it again.

Cooldown notifier — the 5-minute job loaded the entire users table and
did all its DB work on the event loop; the logs caught one run overrunning
its own interval. It now walks the table in bounded id-ordered slices with
a rolling cursor, does the scan and the flag writes on worker threads, and
paces its DMs.

Also from the same logs:
- adsgram_rewards.telegram_id was INTEGER, so every postback from a
  telegram id above 2^31 died with "integer out of range" and cost the
  user their ad reward. Widened to BIGINT, with a migration for existing
  databases.
- finalize_webapp_match checked match status without locking the row, and
  only committed ~100 lines later. Every concurrent caller in that window
  re-finalized the match, which is the burst of "duplicate key ...
  ix_match_scorecards_match_id" errors. The row is now locked before the
  status check; concurrent callers see {"already": True}.

Admin "Send Now" no longer blocks the request for the whole blast — it
waits briefly for an immediate failure, then reports that delivery is
continuing in the background.

+++ models.py
@@ -1538,7 +1538,10 @@ class AdsgramReward(Base):
     """
     __tablename__ = "adsgram_rewards"
     id = Column(Integer, primary_key=True, autoincrement=True)
-    telegram_id = Column(Integer, nullable=False, index=True)
+    # BigInteger, not Integer: Telegram ids have already passed 2^31, and an
+    # INTEGER column rejects them with "integer out of range", failing the
+    # postback and silently costing the user their ad reward.
+    telegram_id = Column(BigInteger, nullable=False, index=True)
     received_at = Column(DateTime, default=datetime.utcnow,
                          nullable=False, index=True)
     consumed_at = Column(DateTime, nullable=True)
